+++ src/setup_scene_description/set_element_arguments_2.c
@@ -6,7 +6,7 @@
 /*   By: aalbrech <aalbrech@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/27 14:21:23 by aalbrech          #+#    #+#             */
-/*   Updated: 2025/04/29 14:05:49 by aalbrech         ###   ########.fr       */
+/*   Updated: 2025/04/29 21:32:47 by aalbrech         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	set_fov(char *element_info, int *fov)
 		error_exit("Camera fov argument must be in the range 0-180");
 }
 
-void	set_diameter_or_height(char *element_info, float *measurement)
+void	set_diameter_or_height(char *element_info, double *measurement)
 {
 	if (ft_strchr(element_info, ','))
 		error_exit("Object diameter or height consists of only one number");
